fix: disable block tests for WC 6.9
